requirements.txt
Flask
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
pytz
gspread
oauth2client
requests
python-dotenv
requests
datetime
gunicorn
openai>=1.0.0
PyMuPDF>=1.23.0
